requirements.txt
glob
numpy
pandas
scipy
imagehash
pillow